refactor: 코드 스플리팅
