+++ .env.sample
@@ -1,0 +1,2 @@
+SLACK_BOT_TOKEN=
+GEULTTO_GITHUB_TOKEN=
