+++ README.md
@@ -1,0 +1,3 @@
+# algo_bot
+
+글또 10기 소모임 "#_오늘도_한문제_풀어또"의 문제 제출 봇입니다.
